Revert "items"
This reverts commit 58552611dcf138cd2f7e0b1c0e4a014ccc02e43c.

+++ Deities Unleashed/Assets/Scripts/EnemyChase.cs
@@ -6,9 +6,9 @@ public class EnemyChase : MonoBehaviour
 {
 
     public Transform Player;
-    [SerializeField] float MoveSpeed = 4;
-    [SerializeField] float MaxDist = 10;
-    [SerializeField] float MinDist = 5;
+    [SerializeField] int MoveSpeed = 4;
+    [SerializeField] int MaxDist = 10;
+    [SerializeField] int MinDist = 5;
 
 
     void Update()
